feat: 视频悬浮预览 + 云端同步UI设置 + 清理音效残留 + 修复import/order
